Build V1 product foundation with Docker, FastAPI backend, and fake login [PL-4]
- FastAPI backend (uv/Python) serving Next.js static export at localhost:8000
- SQLite database initialized fresh on each container start
- Fake login screen at / routing to /dashboard (NDA form)
- Next.js configured for static export; NDA form moved to /dashboard route
- Docker multi-stage build: Node (frontend build) + Python (FastAPI)
- Start/stop scripts for Mac, Linux, and Windows
