+++ Dockerfile
@@ -1,0 +1,22 @@
+# Stage 1: Build Next.js static export
+FROM node:20-alpine AS frontend-builder
+WORKDIR /app
+COPY frontend/package*.json ./
+RUN npm ci --ignore-scripts
+COPY frontend/ ./
+RUN npm run build
+
+# Stage 2: FastAPI backend serving the static frontend
+FROM python:3.12-slim
+WORKDIR /app
+
+RUN pip install --no-cache-dir uv
+
+COPY backend/pyproject.toml backend/uv.lock* ./
+RUN uv sync --frozen --no-dev
+
+COPY backend/main.py ./
+COPY --from=frontend-builder /app/out ./static
+
+EXPOSE 8000
+CMD ["uv", "run", "uvicorn", "main:app", "--host", "0.0.0.0", "--port", "8000"]
