+++ backend/pyproject.toml
@@ -1,0 +1,9 @@
+[project]
+name = "backend"
+version = "0.1.0"
+description = "Add your description here"
+requires-python = ">=3.14"
+dependencies = [
+    "fastapi>=0.134.0",
+    "uvicorn[standard]>=0.41.0",
+]
