+++ docker-compose.yml
@@ -1,0 +1,6 @@
+services:
+  app:
+    build: .
+    ports:
+      - "8000:8000"
+    restart: unless-stopped
