+++ scripts/start-linux.sh
@@ -1,0 +1,5 @@
+#!/bin/bash
+set -e
+cd "$(dirname "$0")/.."
+docker compose up --build -d
+echo "Prelegal running at http://localhost:8000"
