+++ scripts/start-windows.ps1
@@ -1,0 +1,4 @@
+$ErrorActionPreference = "Stop"
+Set-Location (Split-Path $PSScriptRoot)
+docker compose up --build -d
+Write-Host "Prelegal running at http://localhost:8000"
