+++ scripts/stop-linux.sh
@@ -1,0 +1,5 @@
+#!/bin/bash
+set -e
+cd "$(dirname "$0")/.."
+docker compose down
+echo "Prelegal stopped."
