+++ scripts/stop-windows.ps1
@@ -1,0 +1,4 @@
+$ErrorActionPreference = "Stop"
+Set-Location (Split-Path $PSScriptRoot)
+docker compose down
+Write-Host "Prelegal stopped."
